rag-chatbotRAG/requirements.txt
crewai==0.100.1
crewai_tools==0.33.0
pydantic==2.10.6
python-dotenv==1.0.1
